### 2025-08-21 #8849 @editorialbot
[PRE REVIEW]: Condor, a mathematical modeling framework for engineers with a deadline
**Submitting author:** @sixpearls (<a href="http://orcid.org/0000-0001-5602-1888">Benjamin Margolis</a>)
**Repository:** https://github.com/nasa/condor
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.0rc
