### 2025-08-21 #8849 @danielskatz
👋 @sixpearls - Thanks for your submission.  Just to make this slightly more standard for JOSS, could you replace the PDF license with a LICENSE.txt or LICENSE.md file with the same content?

Otherwise, I think this is ready to move forward.
