### 2025-08-21 #8849 @HectorMozo3110
👋 Hi, @danielskatz   I looked into Condor a few days ago and it caught my attention.
If the submission moves to full review, I’d be glad to be considered as one of the reviewers.
