### 2025-08-25 #8859 @editorialbot
[REVIEW]: Condor, a mathematical modeling framework for engineers with a deadline
**Submitting author:** @sixpearls (<a href="http://orcid.org/0000-0001-5602-1888">Benjamin Margolis</a>)
**Repository:** https://github.com/nasa/condor
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.2
