### 2025-08-25 nasa/condor#17 @HectorMozo3110
Dependency Management and Reproducibility Issues
While performing local tests, I set up a clean environment and followed the documented installation process. The core package installs correctly, and all unit tests (pytest) pass. However, when running the example scripts (run_examples.py), several failures occurred:

```
ModuleNotFoundError: No module named 'matplotlib'
```

This error does not mean matplotlib is missing entirely. In fact, the project already declares it under [dependency-groups] docs in pyproject.toml:

```
[dependency-groups]
lint = ["ruff"]
test = ["pytest", "pytest-cov"]
docs = [
  "sphinx",
  "sphinx-gallery",
  "matplotlib",
  "furo",
]
```

The issue is that:
1. **Examples depend on matplotlib**, but the only way to get it is by installing the full docs group (which also installs sphinx and furo).
2. **Tests that produce plots also fail** unless the docs group is installed.
3. **pip ignores [dependency-groups]**, so commands like pip install -e .[test] or pip install -e .[docs] (commonly used in CI and docs) do not actually work. Locally you see warnings like:
```
WARNING: condor does not provide the extra 'test'
```

## Why This Matters (For Users and Collaborators)

- **Reproducibility:** A user or collaborator running examples on a clean environment cannot reproduce results without guessing extra steps.
- **Practicality:** Installing full documentation dependencies just to run run_examples.py is inefficient.
- **Consistency:** The documentation and CI suggest .[test] and .[docs], but pip cannot resolve them.

## Recommended Fix

To maintain both backward compatibility and reproducibility, mirror the dependency groups into [project.optional-dependencies], which pip supports. For example:

```
[project.optional-dependencies]
examples = ["matplotlib"]

test = ["pytest", "pytest-cov"]

docs = [
  "sphinx",
  "sphinx-gallery",
  "matplotlib",  # required by Sphinx-Gallery
  "furo",
]

dev = [
  "pytest", "pytest-cov", "ruff",
  "sphinx", "sphinx-gallery", "matplotlib", "furo",
]
```

## Resulting Workflows

- **Run tests only**
  ```
  pip install -e .[test]
  pytest
  ```

- **Run examples only**
  ```
  pip install -e .[examples]
  python run_examples.py
  ```

- **Build docs**
  ```
  pip install -e .[docs]
  make -C docs html
  ```

- **Full development setup**
  ```
  pip install -e .[dev]
  ```

## Benefit

This keeps the core package lightweight, preserves the modern [dependency-groups] definition, but ensures users, developers, collaborators, and researchers can install exactly what they need. With this change, all examples, tests, and docs become reproducible in a clean environment, ensuring long-term sustainability of the project and aligning fully with JOSS standards.

Note: This observation is part of the ongoing JOSS review (see [openjournals/joss-reviews#8859](https://github.com/openjournals/joss-reviews/issues/8859))
