### 2025-08-25 nasa/condor#18 @HectorMozo3110
Observation (README documentation) (JOSS Review)
@sixpearls
#### README Improvement Suggestions

The current `README.rst` is clear in presenting the project scope, examples of usage, installation from PyPI, and licensing information. However, to fully align with good practices (including JOSS criteria of clarity and reproducibility), I suggest a few targeted improvements that would enhance usability for both new users and reviewers:

#### 1. Execution Instructions
While the README shows installation from PyPI, it would benefit from an explicit section describing how to:

- **Run the test suite** (`pytest`) in a fresh environment.
- **Execute the provided example scripts** (`run_examples.py` or individual files under `docs/examples_src/`).
- **Build the documentation locally** (`make -C docs html`), and optionally as PDF (`make -C docs latexpdf`).

These short command snippets would make it easier for any reader to reproduce the workflows directly without consulting multiple files.

#### 2. Dependencies Clarification
Since some examples and documentation rely on additional dependencies (e.g., plotting, Sphinx), it would be helpful if the README explicitly indicated how to install these optional groups, rather than requiring the user to inspect `pyproject.toml`. This ensures consistency between the declared extras and the documented installation steps. (See also issue [#17](https://github.com/nasa/condor/issues/17) for related discussion.)

#### 3. Citation and Compatibility
Adding a short note on how to cite Condor (DOI, Zenodo entry, or preferred reference) and the officially supported environments (Python ≥3.9, OS compatibility) would complete the README and help academic users.

#### 4. Badges (Optional)
Alongside the existing “Docs / passing / License” badges, a PyPI version badge could be added to highlight the published package version directly.

#### 5. API Overview
While the README includes helpful usage examples, it does not provide an overview of the core API. Adding a short section that lists the main classes, functions, or modules (with a one-line description each) and points to the full documentation would improve clarity for new users and make the package easier to navigate.

---

#### Summary
By including explicit commands for running tests, examples, and documentation, and clarifying optional dependencies, the README will be easier to follow for new users and reviewers. This will strengthen reproducibility and lower the entry barrier for engagement with the project.

**Note:** This finding is also part of the ongoing JOSS review (see [openjournals/joss-reviews#8859](https://github.com/openjournals/joss-reviews/issues/8859)).
