### 2025-08-26 nasa/condor#17 @HectorMozo3110
👋@sixpearls

### Additional Observations on [`pyproject.toml`](https://github.com/nasa/condor/blob/main/pyproject.toml)

Beyond the dependency handling issue previously mentioned, there are additional aspects that could strengthen the robustness and clarity of the package configuration:

1. **License metadata**
   Currently, the license is only included as `license.pdf` and is not declared in the project metadata. This makes it less visible to tools such as PyPI and complicates verification for users.
   **Recommendation:** Declare the license explicitly in `[project]`, for example:
   ```toml
   license = { file = "LICENSE" }
   ```
   This ensures the license information is standardized, automatically displayed on PyPI, and aligned with JOSS requirements for legal clarity and transparency.

**I am still unsure — was leaving the license field empty in the project metadata an intentional decision? If so, I understand; if not, I would suggest addressing this point as mentioned above.**

2. **Explicit `src/` layout declaration**
   The project follows a `src/` layout, which is a good practice, but it is not declared in `pyproject.toml`. Without this, setuptools may not consistently find the package in clean installations (e.g., `pip install .`), leading to potential `import condor` errors.
   **Recommendation:** Add the explicit configuration:
   ```toml
   [tool.setuptools]
   package-dir = {"" = "src"}

   [tool.setuptools.packages.find]
   where = ["src"]
   ```
   This guarantees that the package is always correctly discovered and installed across environments (local, CI/CD, or for new users).

3. **URLs in project metadata (recommended)**
   Adding more links improves discoverability on PyPI and clarity for users:
   ```toml
   [project.urls]
   Homepage = "https://nasa.github.io/condor"
   Documentation = "https://nasa.github.io/condor"
   Repository = "https://github.com/nasa/condor"
   Issues = "https://github.com/nasa/condor/issues"
   ```

4. **Pytest configuration note (optional)**
   The current `addopts` includes `--pdbcls=IPython.terminal.debugger:TerminalPdb`. This can fail if `ipython` is not installed.
   **Recommendation:** Either add `ipython` to the `test` extras or keep this option only for local development.

---
