### 2025-08-29 nasa/condor#19 @amcandio
JOSS review: README example import missing 
Having the import statement on the README examples would be a nice improvement for first-time users. It is just a matter of adding.

```
import condor as co
from condor.backend.operators import exp
```

The exp one was particularly confusing because I initially thought it was `math.exp`.

https://github.com/openjournals/joss-reviews/issues/8859
