### 2025-08-29 nasa/condor#20 @amcandio
JOSS review: casadi dependency and Mac (non-blocking)
Hi! When I ran the example:
```
>>> from condor.backend.operators import exp
>>> class Sellar(co.OptimizationProblem):
...     x = variable(shape=3, lower_bound=0, upper_bound=10)
...     coupling = Coupling(x)
...     y1, y2 = coupling
...
...     objective = x[2]**2 + x[1] + y1 + exp(-y2)
...     constraint(y1 >= 3.16)
...     constraint(24. >= y2)
...
>>> Sellar.set_initial(x=[5,2,1])
... sellar = Sellar()
... print(sellar.objective) # scalar value
... print(sellar.constraint) # field
... print(sellar.coupling.y1) # sub-model element
```

I get:
```
Traceback (most recent call last):
  File "<python-input-17>", line 2, in <module>
    sellar = Sellar()
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/condor/models.py", line 489, in __call__
    return super().__call__(*args, **kwargs)
           ~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/condor/models.py", line 1298, in __init__
    self.implementation = implementation_class(self)
                          ~~~~~~~~~~~~~~~~~~~~^^^^^^
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/condor/implementations/iterative.py", line 23, in __init__
    self.construct(model, **model_instance.options_dict)
    ~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/condor/implementations/iterative.py", line 298, in construct
    self.callback = CasadiNlpsolWarmstart(
                    ~~~~~~~~~~~~~~~~~~~~~^
        primary_function = self.objective_func,
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    ...<11 lines>...
        model_name = model.__name__,
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "<string>", line 16, in __init__
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/condor/solvers/casadi_warmstart_wrapper.py", line 193, in __post_init__
    self.optimizer = casadi.nlpsol(
                     ~~~~~~~~~~~~~^
        f"{self.model_name}_optimizer",
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    ...<2 lines>...
        self.options,
        ^^^^^^^^^^^^^
    )
    ^
  File "/Users/amcandio/condor_pip/condor-pip-dev/lib/python3.13/site-packages/casadi/casadi.py", line 30135, in nlpsol
    return _casadi.nlpsol(*args)
           ~~~~~~~~~~~~~~^^^^^^^
RuntimeError: .../casadi/core/plugin_interface.hpp:317: Plugin 'ipopt' is not found.
```

Seems like a casadi issue with Mac? Is there a known way to solve this?

https://github.com/openjournals/joss-reviews/issues/8859
