### 2025-09-01 nasa/condor#22 @amcandio
JOSS review: include README guidelines to report contributing / issues / seek support
This is a small one. README should briefly mention expected channels to report issues and seek support. For example, pointing to GitHub issues. Moreover, more clarity on how to run the tests would be great for the contributing section.

https://github.com/openjournals/joss-reviews/issues/8859
