### 2025-09-01 #8859 @amcandio
Good stuff! I finished my first pass and created issues with feedback
