### 2025-09-02 nasa/condor#24 @HectorMozo3110
Suggestions to Strengthen the Paper
Hi, 👋@sixpearls

While the current draft is clear and aligns with the JOSS guidelines, I see a few opportunities to strengthen clarity and reproducibility beyond what has already been discussed:

---

#### 1. Comparative Table with Related Tools
The paper mentions several frameworks (Simulink, Modelica, OpenMDAO, Pyomo, etc.), but the comparison is mostly narrative.

**Suggestion:** add a short table contrasting Condor and 2–3 of these tools across criteria such as base language, modularity, solver integration, community adoption.

**Example:**

| Tool     | Base Language | Solver Integration      | Scope               |
|----------|---------------|-------------------------|---------------------|
| Condor   | Python DSL    | CasADi, IPOPT, SUNDIALS | General engineering |
| Modelica | Modelica DSL  | Dymola, OpenModelica    | Physical modeling   |
| OpenMDAO | Python        | Optimizers, DOE         | MDO focus           |

In my opinion, this content could be streamlined by removing the detailed text and keeping it only within the table, which would preserve clarity while helping to keep the paper concise and within the word limit.

---

#### 2. Concise Case Studies
Condor has been used in diverse NASA contexts (aerodynamics, orbital design, propulsion subsystems). Currently these appear only as references.

**Suggestion:** add a short list or table summarizing domain, solver used, and benefit.

**Example:**
- Aerodynamics: VSPAero backend → reduced coding effort in hypersonic studies.
- Trajectory Design: CasADi + SUNDIALS → rapid prototyping of ODE solvers with events.

---

#### 3. Explicit Limitations
The paper highlights Condor’s strengths but does not mention limitations. Adding 1–2 sentences would balance the contribution.

**Examples:**
- Current reliance on CasADi may limit some advanced symbolic workflows.
- Designed for conceptual analysis; not yet optimized for large-scale detailed simulations.

---

#### 4. Visual Architecture Diagram
The “model–solver–backend” description could benefit from a small diagram.

**Suggestion:** a schematic showing *Model (engineer-defined equations) → Solver (IPOPT, SNOPT, SUNDIALS) → Backend (CasADi, JAX)*. This would give readers an immediate grasp of Condor’s modular structure.

---

#### 5. Quickstart Code Snippet
To complement the README examples, the paper itself could include a tiny snippet (3–4 lines) showing how to set up and solve a simple model.

**Example:**

```python
import condor as cd
m = cd.Model()
x = m.add_variable("x")
m.add_equation(x**2 - 4)
sol = m.solve()
print(sol[x])
```

---
#### 6. Future Work
It might also be valuable to include a short paragraph outlining potential future directions for Condor. Even a brief statement of where the software is expected to evolve would help the community better understand its trajectory and could encourage future collaboration or contributions.

---

I believe these improvements would strengthen the paper, but they may require small adjustments to stay within the JOSS word limit. For that reason, my suggestion is to address them mainly with tables and diagrams for comparisons and architecture, and by making small adjustments to a few sentences. This way the manuscript stays concise while improving clarity. It could also help to add a few more references, not only to dependencies already mentioned in the text but also to external comparative sources that situate Condor among related frameworks.

I have seen publications that are around 1000–1100 words, and in my view this paper could also benefit from extending slightly beyond its current word count, as the additional details would significantly strengthen clarity and reproducibility.

**Please feel free to let me know if you have any questions or need further clarification—I’ll be glad to help.**

*Note: This finding is also part of the ongoing JOSS review (see openjournals/joss-reviews#8859).*
