+++ paper/paper.md
@@ -30,7 +30,7 @@ Numerical modeling is an important part of the engineering workflow, providing u
 
 Condor was developed at NASA Ames Research Center's Systems Analysis Office to solve a variety of analysis and design problems in aeronautics [@Margolis2024gascon; @Listgarten2025; @Pham2025; @Park2025; @zelinski2025ttbw; @Margolis2026gascon], orbital trajectory design [@Koehler2024; @Margolis2024techniques; @Margolis2024coopt], and subsystem design [@Margolis2026bwb; @Margolis2026npss]. Condor's modular framework makes it feasible to develop algorithms using existing models as test examples like gradient methods for solutions to ordinary differential equation with events [@Margolis2023sweeping] or uncertain differential equations [@Margolis2026sigma].
 
-A variety of existing libraries work towards unifying solvers and optimization tools under a single interface,however to the best of the author's knowledge, no library provides such a convenient modeling language with a modular framework to leverage existing solvers.
+A variety of existing libraries work towards unifying solvers and optimization tools under a single interface, however to the best of the author's knowledge, no library provides such a convenient modeling language with a modular framework to leverage existing solvers.
 
 After assessing the available tools, we found that nothing met the requirements to allow engineers to have a single, easy-to-use interface to any existing solver and define new models as conceptual analysis demands arose. We built Condor to provide an interface for engineers that used computational tools from the AI/ML community as the "backend" of the framework in the Python programming language. Using Python as the base language offers access to a huge community of practice, especially in scientific computing and numerical methods, where the majority of the AI/ML tools originated. This means that even with a small development team, we could provide features like parallel computing, file interfaces, and more by leveraging the open-source ecosystem.
 
@@ -41,12 +41,12 @@ Condor is a general-purpose engineering-mathematical modeling framework used to 
 
 A major development strategy of Condor is to leverage the existing "competitors" to do the work for us. We use CasADi as our backend, but thanks to Condor's modular architecture, it would be relatively straightforward to swap with alternatives like JAX or Aesara. We also leverage existing best-in-class solvers for optimization (IPOPT, SNOPT, other CasADi-enabled optimizers, sequential least squares SLSQP), differential equations (Runge-Kutta implementations in SciPy, Lawrence Livermore National Lab's SUNDIALS), and efficient array-arithmetic and linear algebra from the backend. Since Condor makes it easy to wrap external solvers into the modeling framework, we have also used NASA-developed solvers like CBAero (for hypersonic aerodynamics and aerothermal modeling), VSPAero (for subsonic and supersonic aerodynamics modeling), and NPSS (for propulsion).
 
-Since Condor uses standard python data structures, we also leverage general purpose scientific tools and libraries. Virtually all utility functions come from the larger python ecosystem:
+Since Condor uses standard Python data structures, we also leverage general purpose scientific tools and libraries. Virtually all utility functions come from the larger Python ecosystem:
 
 - NumPy's built-in file format to read and store model results
 - Compatible with Python parallel processing libraries such as the built-in multiprocessing or the third-party joblib and Dask projects
-- For most low-level operations (creating identity matrix, concatenating vectors,etc), we follow the Python array API standard, promoting user familiarity compatibility with existing libraries potential backends.
-- Compatible with any plotting capability like matplotlib, seaborn, and PyQtGraph.
+- For most low-level operations (creating identity matrix, concatenating vectors, etc), we follow the Python array API standard, promoting user familiarity compatibility with existing libraries potential backends.
+- Compatible with any plotting capability like Matplotlib, seaborn, and PyQtGraph.
 
 
 # References
