### 2025-09-02 6d5d435
rework paper WIP
