+++ paper/paper.md
@@ -30,22 +30,25 @@ Numerical modeling is an important part of the engineering workflow, providing u
 
 Condor was developed at NASA Ames Research Center's Systems Analysis Office to solve a variety of analysis and design problems in aeronautics [@Margolis2024gascon; @Listgarten2025; @Pham2025; @Park2025; @zelinski2025ttbw; @Margolis2026gascon], orbital trajectory design [@Koehler2024; @Margolis2024techniques; @Margolis2024coopt], and subsystem design [@Margolis2026bwb; @Margolis2026npss]. Condor's modular framework makes it feasible to develop algorithms using existing models as test examples like gradient methods for solutions to ordinary differential equation with events [@Margolis2023sweeping] or uncertain differential equations [@Margolis2026sigma].
 
-A variety of existing libraries work towards unifying solvers and optimization tools under a single interface, however to the best of the author's knowledge, no library provides such a convenient modeling language with a modular framework to leverage existing solvers.
+A variety of existing libraries work towards unifying solvers and optimization tools under a single interface.
+After assessing the available tools, we that no existing tool provided an interface that we felt improved the engineering workflow.
+Condor is unique, to the best of the author's knowledge, for providing a mathematically-focused DSL to facilitate rapid prototyping or such a modular architecture to rapidly deploy new or existing solvers for engineering problems. This improves the engineering workflow by providing a single interface to any existing solver supporting definition of new models as conceptual analysis demands arise.
 
-After assessing the available tools, we found that nothing met the requirements to allow engineers to have a single, easy-to-use interface to any existing solver and define new models as conceptual analysis demands arose. We built Condor to provide an interface for engineers that used computational tools from the AI/ML community as the "backend" of the framework in the Python programming language. Using Python as the base language offers access to a huge community of practice, especially in scientific computing and numerical methods, where the majority of the AI/ML tools originated. This means that even with a small development team, we could provide features like parallel computing, file interfaces, and more by leveraging the open-source ecosystem.
+We built Condor to provide an interface for engineers that used computational tools from the AI/ML community as the "backend" of the framework in the Python programming language. 
+Using Python as the base language offers access to a huge community of practice, especially in scientific computing and numerical methods, where the majority of the AI/ML tools originated. This means that even with a small development team, we could provide features like parallel computing, file interfaces, and more by leveraging the open-source ecosystem.
 
 
 # Description
 
 Condor is a general-purpose engineering-mathematical modeling framework used to build conceptual design and analysis capability. It is related to multi-disciplinary analysis (MDA) frameworks (e.g. Simulink, SimuPy, Model Center, NPSS, Modelica), computational tools developed by the Artificial Intelligence/Machine Learning (AI/ML) community (e.g. CasADi, JAX, Aesara, PyTorch, TensorFlow), and specific numerical solvers (e.g., IPOPT/SLSQP for optimization, SUNDIALS for trajectory integration, etc). The goal was to provide a single, easy-to-use interface to the best in-class numerical solvers so engineers can focus on engineering rather than learning solver-specific interfaces or spend their time with intensive coding or algorithm tuning for new models. The computational tools from the AI/ML communities took some steps to address these challenges, but their interfaces did not lend themselves to general engineering approaches to modeling physical systems. Similarly, the existing MDA frameworks did not satisfy the system analysis needs, either because they were too domain focused (e.g. dynamical systems), or because they used their own language so had limited community of practice and could not benefit from recent advancements in best practices and computational techniques.
 
[…]
