### 2025-09-02 nasa/condor#25 @HectorMozo3110
Documentation Feedback (JOSS Review)
👋 @sixpearls

First of all, congratulations on putting together such a rich documentation set for Condor - it already covers tutorials, how-tos, API details, and real examples. That’s a lot of work and it shows. Below are a few suggestions that could help polish things further, making it even clearer and easier to navigate.

---

####  1. Introduction & High-Level Context
It would be nice to start with a very short definition of Condor (2–3 lines) at the beginning, before the section list.
For example:

> “Condor is a Python framework for mathematical modeling and solving algebraic, dynamical, and optimization problems. It provides a math-like domain-specific language (DSL) and a modular model–solver–backend architecture, making it easy to build models and connect them with advanced solvers like CasADi and IPOPT.”

Also, consider adding quick links to the README and the JOSS paper so readers immediately know where to find installation instructions and the scientific background.

---

#### 2. Quickstart Example

- Provide a minimal example (5–8 lines of code) on the landing page, so that someone can install and run Condor in under a minute.

This lowers the entry barrier and immediately showcases the API.

---

#### 3. Solver Output
Right now, the tutorials include the full raw IPOPT logs. This has pros and cons:

- **Pro:** some users will appreciate the full solver trace, since it shows transparency and makes reproducibility straightforward.
- **Con:** for others, especially new users, the amount of output can be overwhelming and distract from the example itself.

**Suggested balance:** in the main docs show only the key parts (main iterations, final objective value, convergence status), and keep the full solver logs available either via links or as optional extended output.

---

#### 4. Consistency & Style
There are a few style details that could be unified:
- Capitalization: always use `Python`, `Condor`, `CasADi`, `SciPy`, `NumPy`, `IPOPT`.
- Fix small typos (e.g. `evaulate → evaluate`, `implementaitons → implementations`, `intitial → initial`).

You can automate checks with `sphinxcontrib-spelling` or similar tools.

---

#### 5. Dependencies & References
It would help to add a short section with references and links to key dependencies:
- [CasADi](https://web.casadi.org/)
- [IPOPT](https://coin-or.github.io/Ipopt/)
- [SciPy](https://docs.scipy.org/doc/scipy/)
- [NumPy](https://numpy.org/doc/stable/)
- [ndsplines](https://ndsplines.readthedocs.io/en/latest/)

This way readers can quickly jump to external docs when they need them.

---

#### 6. Cross-Referencing (Internal API)
In the tutorials, when you use Condor classes, you could add explicit cross-references to your own API.
For example, in reST:

```rst
See also :class:`condor.AlgebraicSystem` and :class:`condor.ExplicitSystem`
```

That way users can easily click through from examples to the detailed API docs.
To make sure none are missing, you can enable `nitpicky = True` in `conf.py`, which will raise warnings if a cross-reference target doesn’t exist.

*(Clarification: this suggestion applies to Condor’s **own API classes** — not external libraries.)*

---

#### 7. Intersphinx (External APIs)
You already have `intersphinx_mapping` set up for Python, NumPy, and SciPy, but plain text mentions won’t automatically become links. To get clickable links, you need to format them with roles, for example:

- `:mod:\`numpy\``
- `:class:\`numpy.ndarray\``
- `:func:\`numpy.linalg.norm\``

In your docstrings, thanks to `napoleon_type_aliases`, you can also make type annotations like `ndarray` resolve to the NumPy docs.
Optionally, setting `autodoc_typehints = "description"` in `conf.py` helps propagate these links automatically.

Together with point 6, this gives you a consistent navigation experience:
- Internal classes (Condor API) link to your own docs.
- External types (NumPy, SciPy, etc.) link to their respective docs.

---

#### 8. Keeping the Docs Clean
A couple of small process tips:
- Enable `sphinx_copybutton` so code blocks have a one-click copy button.
- Keep using `sphinx.ext.intersphinx` (already enabled).
- Use `make linkcheck` regularly to spot broken links.
- Build with `-W --keep-going` to catch warnings systematically without stopping the build.

---

Overall, the documentation is already very strong in technical depth and examples. These suggestions are just to refine clarity, style, and navigation,  making it easier for new users to get started while still keeping the level of detail that advanced users and researchers expect.

**Note:** This observation is part of the ongoing JOSS review (see [openjournals/joss-reviews#8859](https://github.com/openjournals/joss-reviews/issues/8859))
