### 2025-09-02 #8859 @HectorMozo3110
👋 Hi, @sixpearls  first of all congratulations - the documentation and the project already look very complete and well structured.

I’ve now finished my **first review pass**. As part of this, I opened a few issues that you’ve probably already noticed, mainly to make observations and suggestions for improvement:

- [ ] [#25 Documentation Feedback (JOSS Review)](https://github.com/NASA/condor/issues/25)
- [ ] [#24 Suggestions to Strengthen the Paper (JOSS Review)](https://github.com/NASA/condor/issues/24)
- [ ] [#18 Observation (README documentation) (JOSS Review)](https://github.com/NASA/condor/issues/18)
- [ ] [#17 Dependency Management and Reproducibility Issues (JOSS Review)](https://github.com/NASA/condor/issues/17)

We can use those threads to develop the discussion point by point. As we go covering them, I’ll check them off here.

With this, I consider my initial set of suggestions complete, and I’ll be attentive to your responses so we can move towards closing my review.

Thanks again for all the work you’ve put into this,  it’s already in very good shape.
