### 2025-09-03 242918f
attempt at references
