+++ paper/paper.md
@@ -40,16 +40,16 @@ Using Python as the base language offers access to a huge community of practice,
 
 # Description
 
-Condor is a general-purpose engineering-mathematical modeling framework used to build conceptual design and analysis capability. It is related to multi-disciplinary analysis (MDA) frameworks (e.g. Simulink, SimuPy, Model Center, NPSS, Modelica), computational tools developed by the Artificial Intelligence/Machine Learning (AI/ML) community (e.g. CasADi, JAX, Aesara, PyTorch, TensorFlow), and specific numerical solvers (e.g., IPOPT/SLSQP for optimization, SUNDIALS for trajectory integration, etc). The goal was to provide a single, easy-to-use interface to the best in-class numerical solvers so engineers can focus on engineering rather than learning solver-specific interfaces or spend their time with intensive coding or algorithm tuning for new models. The computational tools from the AI/ML communities took some steps to address these challenges, but their interfaces did not lend themselves to general engineering approaches to modeling physical systems. Similarly, the existing MDA frameworks did not satisfy the system analysis needs, either because they were too domain focused (e.g. dynamical systems), or because they used their own language so had limited community of practice and could not benefit from recent advancements in best practices and computational techniques.
+Condor is a general-purpose engineering-mathematical modeling framework used to build conceptual design and analysis capability. It is related to multi-disciplinary analysis (MDA) frameworks (e.g. Simulink [@simulink], SimuPy [@Margolis2017simupy], ModelCenter [@modelcenter], NPSS [@Curlett1995], Modelica [@Fritzson2002]), computational tools developed by the Artificial Intelligence/Machine Learning (AI/ML) community (e.g. CasADi [@Andersson2018], JAX [@Bradbury2018], Aesara [@Willard2023], PyTorch [@Paszke2017], TensorFlow [@MartinAbadi2015]), and specific numerical solvers (e.g., IPOPT [@Waechter2005] / SLSQP [@Kraft1994] for optimization, SUNDIALS [@Hindmarsh2005] for trajectory integration, etc). The goal was to provide a single, easy-to-use interface to the best in-class numerical solvers so engineers can focus on engineering rather than learning solver-specific interfaces or spend their time with intensive coding or algorithm tuning for new models. The computational tools from the AI/ML communities took some steps to address these challenges, but their interfaces did not lend themselves to general engineering approaches to modeling physical systems. Similarly, the existing MDA frameworks did not satisfy the system analysis needs, either because they were too domain focused (e.g. dynamical systems), or because they used their own language so had limited community of practice and could not benefit from recent advancements in best practices and computational techniques.
 
-Condor was designed and built following modern software development best-practices. We leveraged CasADi as our backend, but thanks to Condor's modular architecture, changing backends (e.g., to JAX or Aesara) would require providing a few interface files for the backend-shim infrastructure. We also leverage existing best-in-class solvers for optimization (IPOPT, SNOPT, other CasADi-enabled optimizers, sequential least squares SLSQP), differential equations (Runge-Kutta implementations in SciPy, Lawrence Livermore National Lab's SUNDIALS), and efficient array-arithmetic and linear algebra from the backend. Since Condor makes it easy to wrap external solvers into the modeling framework, we have also used NASA-developed solvers like CBAero (for hypersonic aerodynamics and aerothermal modeling), VSPAero (for subsonic and supersonic aerodynamics modeling), and NPSS (for propulsion).
[…]
