+++ paper/references.bib
@@ -121,4 +121,244 @@
   publisher = {Springer Science and Business Media LLC},
 }
 
+@Article{Andersson2018,
+  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
+            and James B Rawlings and Moritz Diehl},
+  Title = {{CasADi} -- {A} software framework for nonlinear optimization
+           and optimal control},
+  Journal = {Mathematical Programming Computation},
+  Year = {2018},
+}
+
+@Software{Bradbury2018,
+  author  = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and Skye Wanderman-Milne},
+  title   = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url     = {http://github.com/google/jax},
+  version = {0.1.46},
+  year    = {2018},
+}
+
+@Article{Waechter2005,
+  author    = {Andreas Wächter and Lorenz T. Biegler},
+  journal   = {Mathematical Programming},
+  title     = {On the Implementation of an Interior-Point Filter Line-Search Algorithm for Large-Scale Nonlinear Programming},
+  year      = {2005},
+  month     = {apr},
+  number    = {1},
+  volume    = {106},
+  doi       = {10.1007/s10107-004-0559-y},
+  publisher = {Springer Science and Business Media {LLC}},
+}
+
+@Software{simulink,
+  author = {{The MathWorks, Inc.}},
+  title  = {Simulink - Simulation and model-based design},
+  url    = {https://www.mathworks.com/products/simulink.html},
+}
+
+@Article{Margolis2017simupy,
+  author  = {Margolis, Benjamin WL},
+  journal = {J. Open Source Software},
+  title   = {{SimuPy}: A {Python} framework for modeling and simulating dynamical systems},
+  year    = {2017},
+  number  = {17},
+  pages   = {396},
+  volume  = {2},
+  doi     = {10.21105/joss.00396},
+}
+
+@Software{modelcenter,
+  author = {{Phoenix Integration}},
+  title  = {ModelCenter},
+  url    = {https://www.ansys.com/products/connect/ansys-modelcenter?intcid=phoenixintreferral},
+}
+
+@TechReport{Curlett1995,
+  author = {Curlett, Brian P and Felder, James L},
+  title  = {Object-oriented approach for gas turbine engine simulation},
+  year   = {1995},
+  number = {106970},
+  type   = {techreport},
+  school = {NASA},
+}
+
+@InProceedings{Fritzson2002,
+  author    = {Fritzson, P. and Bunus, P.},
+  booktitle = {Proceedings 35th Annual Simulation Symposium. SS 2002},
+  title     = {Modelica - a general object-oriented language for continuous and discrete-event system modeling and simulation},
+  year      = {2002},
+  pages     = {365-380},
+  doi       = {10.1109/SIMSYM.2002.1000174},
+  keywords  = {Object oriented modeling;Discrete event systems;Electrical capacitance tomography;Equations;Biological system modeling;Mathematical model;Computer simulation;Thermodynamics;Application software;Computational modeling},
+}
+
+@Software{Willard2023,
+  abstract     = {Aesara is a Python library for defining, optimizing, and efficiently evaluating mathematical expressions involving multi-dimensional arrays.},
+  address      = {{GitHub}},
[…]
