### 2025-09-03 nasa/condor#20 @sixpearls
That's a good point. It looks like the issue is marked for Casadi's 3.8 milestone and it looks like some progress is being made today. We will monitor the situation and depending on the anticipated timelines for releases of Condor and Casadi, add some form of documentation for addressing this.
