### 2025-09-04 31a9fa8
skip parallel processing example during docs build on mac
