### 2025-09-04 0ddb4e0
Add dev instructions to README
