### 2025-09-04 99644ed
Add pytest config per sp-repo-review
