### 2025-09-04 c133e38
Add more project URLs
