### 2025-09-05 aa4a9c6
Add dev dependency group and add instructions for dev install
