+++ README.rst
@@ -83,7 +83,7 @@ install with pip:
 
 .. code:: bash
 
-   $ pip install condor
+   pip install condor
 
 Contributing
 ------------
@@ -111,7 +111,9 @@ Development
 ^^^^^^^^^^^
 
 Clone the repository and install it in editable mode along with the ``dev``
-dependency group in a virtual environment::
+dependency group in a virtual environment:
+
+.. code:: bash
 
     pip install --upgrade pip
     pip install -e . --group dev
@@ -119,7 +121,9 @@ dependency group in a virtual environment::
 A `pre-commit <https://pre-commit.com/>`_ configuration is available to
 automatically format and fix linting issues when creating a commit. Install
 pre-commit either in the virtual environment created above or as a system-wide
-tool (e.g. pipx or uv tool), then install the hooks::
+tool (e.g. pipx or uv tool), then install the hooks:
+
+.. code:: bash
 
     pre-commit install
 
@@ -134,16 +138,37 @@ To run the tests, run::
 License
 -------
 
-This software is released under the `NASA Open Source Agreement Version 1.3 <https://github.com/nasa/condor/raw/main/license.pdf>`_.
+This software is released under the `NASA Open Source Agreement Version 1.3
+<https://github.com/nasa/condor/raw/main/license.pdf>`_.
 
 Notices
 -------
 
-Copyright © 2024 United States Government as represented by the Administrator of the National Aeronautics and Space Administration.  All Rights Reserved.
+Copyright © 2024 United States Government as represented by the Administrator of
+the National Aeronautics and Space Administration.  All Rights Reserved.
 
 Disclaimers
 -----------
 
[…]
