+++ .github/workflows/test.yml
@@ -40,8 +40,25 @@ jobs:
     - name: Run tests
       run: pytest --cov=src
 
-    - name: Install example dependencies
-      run: python -m pip install --group docs
+  examples:
+    name: Run examples
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+        os: [ubuntu-latest, windows-latest, macos-latest]
+
+    steps:
+    - uses: actions/checkout@v4
+
+    - uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install dependencies
+      run: |
+        python -m pip install .[examples]
 
     - name: Run examples
       env:
