### 2025-09-05 nasa/condor#24 @HectorMozo3110
I consider that the paper is the visible face of the software. While the submission clearly meets the minimum requirements outlined in the JOSS guidelines, in this paper I find that the manuscript could benefit from greater clarity.

Zenodo normally offers a DOI for GitHub repositories, including the code, documentation, and all related materials. However, in my criterion the paper is important—this is precisely what differentiates JOSS, since the paper is often the first thing readers and users will see, and it should provide a clear, concise, and accessible academic entry point to the project.

In this specific case, I feel the paper could be clearer in some aspects, and that is why I previously provided suggestions to help strengthen clarity and organization. As it stands, I cannot issue an acceptance judgment, since I believe new readers could find it somewhat difficult to follow. Therefore, I will leave the decision on this aspect to the editor, @danielskatz.
