### 2025-09-05 nasa/condor#18 @HectorMozo3110
Hi 👋 @sixpearls,

Thank you for addressing these points. Some of my suggestions were optional, and I see that the updates and clarifications you’ve provided in the README are sufficient and aligned with JOSS practices. I consider this issue resolved, so I will close this point.
