### 2025-09-05 nasa/condor#24 @sixpearls
[From the blog post announcing JOSS](https://www.arfon.org/announcing-the-journal-of-open-source-software#-oh-cool-youre-going-to-publish-a-bunch-of-crappy-papers)

> JOSS “papers” are essentially just abstracts that point to a software repository.

The paper is not intended to be more than a brief, academically-citable, reference to the software.

> I feel the paper could be clearer in some aspects

The recommendations in this list add exposition and documentation; the other JOSS reviewer did provide feedback to improve clarity which we implemented
