### 2025-09-05 nasa/condor#24 @danielskatz
👋 @HectorMozo3110 and @sixpearls - I generally agree with @sixpearls here. JOSS papers are not meant to be full descriptions of the software, but brief (<1000 word) placeholders, that point to content elsewhere (repo, docs) in many cases.  We generally don't expect that people will find the software via the paper, but rather, that the paper enables the developers to get academic publication credit for the work they did in the software project.  See https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper for what should be in a JOSS paper.

@HectorMozo3110 - if you think there is important information missing from the project that is part of the review criteria, it's perfectly reasonable to ask the submitters for it.

Let me know if this makes sense or if more discussion is needed.

Thanks both!
