### 2025-09-05 nasa/condor#24 @HectorMozo3110
Thank you @danielskatz, that clarification helps. I was reading the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), and I may have misinterpreted parts of it. In my criterion, I felt there were areas where the paper could be clearer, so I raised them only as suggestions, not strict requirements.

I understand that a JOSS paper has length limits and is not meant to be extensive, but it should still be clear — and that was the intent behind my suggestions. I opened this thread mainly to clarify that point, since I wasn’t fully sure.

Sincerely, when reading it I don’t feel it is entirely clear, though I agree it meets the JOSS requirements. I will therefore close this point, but I just wanted confirmation, since in my view the paper could still be clearer in some aspects.

- **Summary for non-specialists**: currently includes technical jargon (e.g., *metaprogramming, DSL, modular model–solver–backend*) that may not be immediately accessible.
- **Statement of need and context of related work**: related tools are listed, but the differences vs. alternatives.
- **References**: several are dense and use abbreviations for venues; full venue names and less clustering would improve readability.

**Reference**
