### 2025-09-05 nasa/condor#24 @danielskatz
@HectorMozo3110 - your suggestions above seem reasonable to me.
