### 2025-09-06 ffe23f2
add copybutton to sphinx conf
