### 2025-09-06 ac32722
add casadi link to docs
