### 2025-09-06 nasa/condor#25 @sixpearls
> 1. Introduction & High-Level Context
> 2. Quickstart Example

The introduction from the README has been included in the documentation index page which includes an introduction, high-level context, and a quick start example.

> 3. Solver Output

We like to use the full solver output in the tutorials because (1) it is the default (2) it shows progress feedback

> 4. Consistency & Style

We ran the docs through spellchecker and searched for the library names to make corrections. Please let us know if you identify any additional corrections

> 5. Dependencies & References

I am not familiar with indices of external references in docs before. I added a link to CasADi where it was mentioned by name in the design document. We will keep this suggestion mind when adding future documentation.

> 6. Cross-Referencing (Internal API)

We have done this where we thought it was appropriate and enabled search in the docs which should make it easy enough to find the necessary information

> 7. Intersphinx (External APIs)

I am not seeing anywhere we are mentioning external modules in the docs. We will keep this suggestion mind when adding future documentation.

> 8. Keeping the Docs Clean

Of the suggestions, only adding `sphinx_copybutton` was actionable, which we did
