### 2025-09-06 nasa/condor#25 @HectorMozo3110
👋 @sixpearls,

Thanks for the updates — I’m happy with these changes and clarifications. I understand the project is under constant development, but for the purpose of finalizing my review, these updates are sufficient. I’ll consider this point closed.
