### 2025-09-07 nasa/condor#22 @amcandio
That's great, closing this one!
