### 2025-09-07 #8859 @amcandio
We have confirmation that the casadi issue is going to be resolved in the next release so I don't believe it's a blocker for the paper.

Thank you for building this. I recommend the paper to be approved for JOSS.
