### 2025-09-08 #8859 @HectorMozo3110
👋 @sixpearls,

Thanks again for the clarifications and for your work on Condor!. All the issues raised during the review have been resolved, and I’m satisfied with the current state of the software.

With this, I consider my review complete. I am happy to recommend acceptance of this submission for JOSS.

@danielskatz — I confirm that my review is now finished.
