### 2025-09-08 #8859 @danielskatz
@sixpearls - I've also suggested a few wording changes in https://github.com/nasa/condor/pull/27 - please merge this, or let me know what you disagree with.
