### 2025-09-09 18635f6
Update paper.md (#27)
