+++ paper/paper.md
@@ -32,7 +32,7 @@ Condor was developed at NASA Ames Research Center's Systems Analysis Office to s
 
 A variety of existing libraries work towards unifying solvers and optimization tools under a single interface.
 After assessing the available tools, we found that no existing tool provided an interface that we felt improved the engineering workflow.
-Condor is unique, to the best of the author's knowledge, for providing a mathematically-focused DSL to facilitate rapid prototyping of engineering models. Condor's modular architecture supports rapid deployment new or existing solvers for engineering problems. This improves the engineering workflow by providing a single interface to any existing solver supporting definition of new models as conceptual analysis demands arise.
+Condor is unique, to the best of the authors' knowledge, for providing a mathematically-focused DSL to facilitate rapid prototyping of engineering models. Condor's modular architecture supports rapid deployment new or existing solvers for engineering problems. This improves the engineering workflow by providing a single interface to any existing solver supporting definition of new models as conceptual analysis demands arise.
 
 We built Condor to provide an interface for engineers that used computational tools from the AI/ML community as the "backend" of the framework in the Python programming language. 
 Using Python as the base language offers access to a huge community of practice, especially in scientific computing and numerical methods, where the majority of the AI/ML tools originated. This means that even with a small development team, we could provide features like parallel computing, file interfaces, and more by leveraging the open-source ecosystem.
@@ -40,7 +40,7 @@ Using Python as the base language offers access to a huge community of practice,
 
 # Description
 
-Condor is a general-purpose engineering-mathematical modeling framework used to build conceptual design and analysis capability. It is related to multi-disciplinary analysis (MDA) frameworks (e.g. Simulink [@simulink], SimuPy [@Margolis2017simupy], ModelCenter [@modelcenter], NPSS [@Curlett1995], Modelica [@Fritzson2002]), computational tools developed by the Artificial Intelligence/Machine Learning (AI/ML) community (e.g. CasADi [@Andersson2018], JAX [@Bradbury2018], Aesara [@Willard2023], PyTorch [@Paszke2017], TensorFlow [@MartinAbadi2015]), and specific numerical solvers (e.g., IPOPT [@Waechter2005] / SLSQP [@Kraft1994] for optimization, SUNDIALS [@Hindmarsh2005] for trajectory integration, etc). The goal was to provide a single, easy-to-use interface to the best in-class numerical solvers so engineers can focus on engineering rather than learning solver-specific interfaces or spend their time with intensive coding or algorithm tuning for new models. The computational tools from the AI/ML communities took some steps to address these challenges, but their interfaces did not lend themselves to general engineering approaches to modeling physical systems. Similarly, the existing MDA frameworks did not satisfy the system analysis needs, either because they were too domain focused (e.g. dynamical systems), or because they used their own language so had limited community of practice and could not benefit from recent advancements in best practices and computational techniques.
[…]
