### 2025-09-10 #8859 @danielskatz
@sixpearls - if you can put the changes in a PR, that might be easiest for everyone to look at
