### 2025-09-10 ddbbc8d
typos and wording
