### 2025-09-10 ee26873
update casadi ref
