+++ paper/references.bib
@@ -121,15 +121,6 @@
   publisher = {Springer Science and Business Media LLC},
 }
 
-@Article{Andersson2018,
-  Author = {Joel A E Andersson and Joris Gillis and Greg Horn
-            and James B Rawlings and Moritz Diehl},
-  Title = {{CasADi} -- {A} software framework for nonlinear optimization
-           and optimal control},
-  Journal = {Mathematical Programming Computation},
-  Year = {2018},
-}
-
 @Software{Bradbury2018,
   author  = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and Skye Wanderman-Milne},
   title   = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
@@ -359,6 +350,20 @@
   doi       = {10.1038/s41586-020-2649-2},
   publisher = {Springer Science and Business Media {LLC}},
   url       = {https://doi.org/10.1038/s41586-020-2649-2},
+}
+
+@Article{Andersson2018,
+  author    = {Joel A E Andersson and Joris Gillis and Greg Horn and James B Rawlings and Moritz Diehl},
+  journal   = {Mathematical Programming Computation},
+  title     = {{CasADi} -- {A} software framework for nonlinear optimization and optimal control},
+  year      = {2018},
+  issn      = {1867-2957},
+  month     = jul,
+  number    = {1},
+  pages     = {1--36},
+  volume    = {11},
+  doi       = {10.1007/s12532-018-0139-4},
+  publisher = {Springer Science and Business Media LLC},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
