### 2025-09-10 #8859 @HectorMozo3110
@sixpearls  Looks good to me. That said, could you go ahead and merge [PR #26](https://github.com/nasa/condor/pull/26) so we can see the license updates and other changes reflected?
