### 2025-09-15 #8859 @danielskatz
👋 @sixpearls - Can you take a look at this proof and make sure everything looks ok to you, as I will also do, before we move on to published the submission?
