### 2025-09-16 #8859 @danielskatz
Congratulations to @sixpearls (Benjamin Margolis) and co-author on your JOSS publication!!

And thanks to @HectorMozo3110 & @amcandio for reviewing!
JOSS completely depends on volunteers, and we couldn't do this without you
